+++ device-discovery/tests/custom_drivers/brocade_netiron/__init__.py
@@ -1,0 +1,1 @@
+"""Unit tests for the custom_napalm.brocade_netiron NetIron driver."""
