+++ device-discovery/tests/custom_drivers/brocade_netiron/mock_data/test_get_config/normal/show_running-config.txt
