+++ device-discovery/tests/custom_drivers/brocade_netiron/mock_data/test_get_facts/normal/show_interfaces_brief.txt
