+++ device-discovery/tests/custom_drivers/brocade_netiron/mock_data/test_get_facts/normal/show_version.txt
