+++ device-discovery/tests/custom_drivers/brocade_netiron/mock_data/test_get_interfaces/link_down/show_interfaces.txt
