--- device-discovery/tests/custom_drivers/netiron/__init__.py
@@ -1,1 +1,0 @@
-"""Unit tests for the custom_napalm.netiron NetIron driver."""
